refactor: Modify component names
