+++ client/src/composition/StudyPages/StudyPageForm/CompleteBlock.tsx
@@ -1,25 +1,33 @@
 import React, { Fragment } from 'react';
 import styled from 'styled-components';
-import PassWordGenerator from './PassWordGenerator';
 
-interface FormProps {
-  setStep: (number) => void;
-}
-
-function Form2({ setStep }: FormProps) {
+function CompleteBlock() {
   return (
     <Fragment>
       <Layout>
-        <Title>스터디룸 비밀번호를 설정해주세요</Title>
-        <PassWordGenerator />
-        <SubmitButton onClick={() => setStep(3)}>완료</SubmitButton>
+        <Title>
+          🎉 축하합니다 🎉
+          <br />
+          <br />
+          스터디 이름 스터디가 생성되었습니다.
+        </Title>
+        <StudyLink>https://study-around.com/la24lk1lk</StudyLink>
+        <SubmitButton>홈으로 가기</SubmitButton>
       </Layout>
     </Fragment>
   );
 }
 
-export default Form2;
+export default CompleteBlock;
 
+const StudyLink = styled.div`
+  text-decoration: underline;
+  margin-top: 30px;
+  text-align: center;
+  font-size: 18px;
+  font-weight: bold;
+  color: #808080;
+`;
 const SubmitButton = styled.div`
   cursor: pointer;
   margin-top: 100px;
@@ -35,13 +43,12 @@ const SubmitButton = styled.div`
   border-radius: 5px;
   border: none;
 `;
-
 const Title = styled.div`
-  font-size: 20px;
+  text-align: center;
+  font-size: 24px;
   font-weight: bold;
-  color: #737373;
+  color: #000;
 `;
-
 const Layout = styled.div`
   display: flex;
   justify-content: center;
